Move launching code that depends on JDT/Cucumber-JVM to a separate plugin
Create a new plugin for this code as it depends on JDT/Cucumber-JVM to
do the launching. These dependencies tie the editor to a particlar cucumber
implementation which is undesirable for re-use of the editor component.

Minor oddity: Using a cross-plugin reference for the icon did not work out
when launching the plugin into an Eclipse 3 build. Possibly not supported
by the corresponding UI that shows the launch configuration. So the
cukes.gif image is copied from the cucumber.eclipse.editor plugin.