Pass the actual feature file to the IStepDefinitions instance
This can be useful in cases where the lookup of step implementations
wants to consider the directory of the feature file for finding steps
instead of going through a complete project. In particular to avoid reporting
steps that are not usable from a certain feature file due to them being
in unrelated directories.