Add JDT steps builder for keeping steps in sync with Gherkin model
Also include a nature, and a handler for adding the nature under
the project Convert To menu. The Gherkin model is notified of
changes to the steps by a listener registered on IStepDefinitions.

Modify content assist classes to use the live step definitions via
the IStepProvider.

Fixes #122 using the approach outlined there.