Add support for validation plugins from the project
Currently it is only possible to validate the glue on execution but it
some cases it would be useful to have some prevalidation.

This adds support for specify validation plugins in the glue code that
are executed as part of the parsing of document and show up as errors
immediately on save.