Fix wrong CucumberJavaBackendProperties#setEnabled method
CucumberJavaBackendProperties#setEnabled does not do what it claims and
instead of just fetches the current value as a result of C&P error.

THis now fixes the method to correctly update the value, beside that it
exports all packages (as internal) to allow access for other plugins
that want to change java settings programmatically.